Align tensor dtypes with GPT-style workloads
Drop the extra float/index variants so the core tensor code matches the types we actually need for transformer work. This keeps the implementation smaller while making token ids and low-precision execution line up with PyTorch conventions.

Move index tensors and targets to i64 for embedding and loss compatibility, and keep floating-point support focused on f32 and native f16 across CPU and MPS paths. The result is a simpler dtype surface that is easier to reason about while still covering the practical training and inference cases DEERS is aiming to teach.

+++ examples/mnist_train.rs
@@ -106,7 +106,7 @@ fn evaluate(model: &impl Module, images: &Tensor, labels: &Tensor, batch_size: u
     for batch_idx in 0..num_batches {
         let start = batch_idx * batch_size;
         let batch_images = images.narrow(0, start, batch_size);
-        let batch_labels: Vec<u32> = labels.narrow(0, start, batch_size).to_vec().unwrap();
+        let batch_labels: Vec<i64> = labels.narrow(0, start, batch_size).to_vec().unwrap();
 
         let logits = model.forward(&batch_images).unwrap();
         let logit_vals: Vec<f32> = logits.to_vec().unwrap();
